requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
blinker==1.9.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudpathlib==0.23.0
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.3
confection==0.1.5
contourpy==1.3.2
cryptography==46.0.3
cycler==0.12.1
cymem==2.0.11
debugpy==1.8.17
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.120.0
filelock==3.20.0
Flask==3.1.2
fonttools==4.60.1
fsspec==2025.9.0
h11==0.16.0
huggingface-hub==0.36.0
idna==3.11
imageio==2.37.0
ipykernel==7.0.1
ipython==8.37.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.9
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
lime==0.2.0.1
llvmlite==0.45.1
lxml==6.0.2
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.13
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.2
numba==0.62.1
numpy==2.2.6
packaging==25.0
pandas==2.3.3
parso==0.8.5
pdfminer.six==20250506
pillow==12.0.0
platformdirs==4.5.0
preshed==3.0.10
prompt_toolkit==3.0.52
psutil==7.1.1
pure_eval==0.2.3
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
regex==2025.10.23
requests==2.32.5
rich==14.2.0
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.1.2
shap==0.49.1
shellingham==1.5.4
six==1.17.0
slicer==0.0.8
smart_open==7.4.1
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
starlette==0.48.0
sympy==1.14.0
thinc==8.3.6
threadpoolctl==3.6.0
tifffile==2025.5.10
tokenizers==0.22.1
torch==2.9.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
wasabi==1.1.3
wcwidth==0.2.14
weasel==0.4.1
Werkzeug==3.1.3
wrapt==2.0.0
